+++ test/core.spec.js
@@ -4,6 +4,7 @@ const complexJSON = require('./examples/transformValues/complex.json');
 
 describe('core', () => {
   let NODE_MASTER_KEY = '8aa93853b3ff01c5b5447529a9c33cb9';
+  const SHORT_NODE_MASTER_KEY = '12345678';
   let credentials = { key: 'value' };
   let credentialsString = JSON.stringify(credentials);
   test('encrypt', async () => {
@@ -12,7 +13,7 @@ describe('core', () => {
   });
 
   test('encrypt with any key length', async () => {
-    const result = await core.encrypt('12345678', credentialsString);
+    const result = await core.encrypt(SHORT_NODE_MASTER_KEY, credentialsString);
     expect(result).validEncrypted();
   });
 
@@ -20,6 +21,14 @@ describe('core', () => {
     const [result, iv] = await core.decrypt(
       NODE_MASTER_KEY,
       'dwAhexc3PhUrX9i4gutpy6Hb8endKm7hMCQALPspYEc=--84X822lxzoPbO9Jh2knEGA=='
+    );
+    expect(JSON.parse(result)).toEqual(credentials);
+  });
+
+  test('decrypt  with any key length', async () => {
+    const [result, iv] = await core.decrypt(
+      SHORT_NODE_MASTER_KEY,
+      'YhSWFEmk0OOG1qQDmjkEjg==--hfHcXE55MQ0bDOHho2SLag=='
     );
     expect(JSON.parse(result)).toEqual(credentials);
   });
@@ -32,5 +41,10 @@ describe('core', () => {
   test('transformValues async', async () => {
     const result = await core.transformValues(complexJSON, async (value) => `${value} MOCK`);
     expect(result).toMatchSnapshot();
+  });
+
+  test('newKey', () => {
+    const result = core.newKey();
+    expect(result).toHaveLength(32);
   });
 });